+++ src/filemanager/file.h
@@ -5,8 +5,7 @@
 #ifndef MC__FILE_H
 #define MC__FILE_H
 
-#include <sys/types.h>          /* off_t */
-#include <sys/time.h>
+#include <inttypes.h>            /* off_t, uintmax_t */
 
 #include "lib/global.h"
 #include "lib/widget.h"
@@ -56,9 +55,8 @@ FileProgressStatus file_error (const char *format, const char *file);
 /* return value is FILE_CONT or FILE_ABORT */
 FileProgressStatus compute_dir_size (const char *dirname, const void *ui,
                                      compute_dir_size_callback cback,
-                                     off_t * ret_marked, double *ret_total,
+                                     off_t * ret_marked, uintmax_t *ret_total,
                                      gboolean compute_symlinks);
-
 
 ComputeDirSizeUI *compute_dir_size_create_ui (void);
 void compute_dir_size_destroy_ui (ComputeDirSizeUI * ui);
